Updated App component tests to reflect name changes and reorganization
